Delete README.md
Simple Inventory Management is a simple system that can register, view product items and sales. Also, the system is useful as a model of how to construct a simple inventory system within Python. This simple project has text file to store the current inventory details. Also, this is a simple json based project which is easy to understand and use.